mcp/requirements.txt
streamlit>=1.28.0
ollama>=0.1.7
fastmcp>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.25.0
uvicorn>=0.24.0
dnspython>=2.4.0
